portal/pyproject.toml
[tool.poetry]
name = "portal"
version = "0.1.0"
description = "LDSA Portal for running Starters Academy"
authors = [
    "Hugo Castilho <hugo.castilho@lisbondatascience.org>",
    "Miguel Dias <miguel.dias@lisbondatascience.org>",
    "Yash Pandya <yash.pandya@lisbondatascience.org>",
    "Gustavo Fonseca <gustavo.fonseca@lisbondatascience.org>",
    "João Gomes <joao.gomes@lisbondatascience.org>",
    "Carlos Regis <carlos.regis@lisbondatascience.org>",
    "Mária Hanulová <maria.hanulova@lisbondatascience.org>",
]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pytz = "^2023.3"
python-slugify = "^8.0.1"
pillow = "^10.0.0"
rcssmin = "^1.1.1"
argon2-cffi = "^23.1.0"
redis = "^5.0.0"
cryptography = "^41.0.3"
pyyaml = "^6.0.1"
python-dateutil = "^2.8.2"
nbconvert = "^7.7.4"
nbformat = "^5.9.2"
numpy = "^1.25.2"
pandas = "^2.0.3"
scikit-learn = "^1.3.0"
python-dotenv = "^1.0.0"
crispy-bootstrap4 = "^2022.1"
# Django Dependencies
django = "^4.2.4"
django-anymail = "^10.1"
django-allauth = "^0.55.0"
django-compressor = "^4.4"
django-constance = {extras = ["database"], version = "^3.1.0"}
django-crispy-forms = "^2.0"
django-environ = "^0.10.0"
django-model-utils = "^4.3.1"
django-redis = "^5.3.0"
djangorestframework = "^3.14.0"

[tool.poetry.group.prod.dependencies]
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.7"
collectfast = "^2.2.0"
sentry-sdk = "^1.29.2"
boto3 = "^1.28.35"
django-anymail = {extras = ["amazon-ses"], version = "^10.1"}

[tool.poetry.group.dev.dependencies]
werkzeug = "^2.3.7"
ipdb = "^0.13.13"
psycopg2-binary = "^2.9.7"
mypy = "^1.5.1"
pytest = "^7.4.0"
pytest-sugar = "^0.9.7"
coverage = "^7.3.0"
factory-boy = "^3.3.0"
django-debug-toolbar = "^4.2.0"
django-extensions = "^3.2.3"
django-coverage-plugin = "^3.1.0"
pytest-django = "^4.5.2"
pre-commit = "^3.4.0"
black = "^23.7.0"
ruff = "^0.0.290"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[tool.mypy.overrides]
module = ["*.migrations.*"]
ignore_errors = true

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
(
      docs
    | migrations
    | static/CACHE/
)
'''

[tool.ruff]
cache-dir = "~/.cache/ruff"
fix = false
target-version = "py310"
show-fixes = true
show-source = true
select = ["ALL"]
ignore = [
  "ANN",    # Type hints related, handled by MyPy
  "COM812", # Trailing comma missing, handled by Black
  "D211",   # No blank lines allowed before class docstring
  "D213",   # Multi-line docstring summary should start at the second line
  "E501",   # line too long, handled by Black
  "RUF100", # Unused noqa directive
]
extend-exclude = [
  "**/migrations/*",
  "**/static/CACHE/*",
  "docs",
]

[tool.ruff.pep8-naming]
ignore-names = [
    'foo',
    'bar',
]

[tool.pytest.ini_options]
norecursedirs = "node_modules"
addopts = '--ds=config.settings.test --ignore=portal/applications --ignore=portal/candidate --ignore=portal/selection --ignore=portal/staff'
testpaths = ['portal']
